[UPDATE] Move ColorCard to card package.
